refactor(forms): move value accessor tests into own spec (#18356)
PR Close #18356
