+++ packages/forms/test/reactive_integration_spec.ts
@@ -6,14 +6,15 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Component, Directive, EventEmitter, Input, Output, Type, forwardRef} from '@angular/core';
+import {Component, Directive, Input, Type, forwardRef} from '@angular/core';
 import {ComponentFixture, TestBed, fakeAsync, tick} from '@angular/core/testing';
-import {AbstractControl, AsyncValidator, AsyncValidatorFn, COMPOSITION_BUFFER_MODE, ControlValueAccessor, FormArray, FormControl, FormGroup, FormGroupDirective, FormsModule, NG_ASYNC_VALIDATORS, NG_VALIDATORS, NG_VALUE_ACCESSOR, NgControl, ReactiveFormsModule, Validators} from '@angular/forms';
+import {AbstractControl, AsyncValidator, AsyncValidatorFn, COMPOSITION_BUFFER_MODE, FormArray, FormControl, FormGroup, FormGroupDirective, FormsModule, NG_ASYNC_VALIDATORS, NG_VALIDATORS, ReactiveFormsModule, Validators} from '@angular/forms';
 import {By} from '@angular/platform-browser/src/dom/debug/by';
 import {getDOM} from '@angular/platform-browser/src/dom/dom_adapter';
 import {dispatchEvent} from '@angular/platform-browser/testing/src/browser_util';
 import {timer} from 'rxjs/observable/timer';
 import {_do} from 'rxjs/operator/do';
+import {MyInput, MyInputForm} from './value_accessor_integration_spec';
 
 export function main() {
   describe('reactive forms integration tests', () => {
@@ -75,7 +76,7 @@ export function main() {
 
     });
 
-    describe('rebound form groups', () => {
+    describe('re-bound form groups', () => {
 
       it('should update DOM elements initially', () => {
         const fixture = initTest(FormGroupComp);
@@ -108,24 +109,6 @@ export function main() {
         newForm.setValue({login: 'Carson'});
         fixture.detectChanges();
         expect(input.nativeElement.value).toEqual('Carson');
-      });
-
-      it('should work with radio buttons when reusing control', () => {
-        const fixture = initTest(FormControlRadioButtons);
-        const food = new FormControl('chicken');
-        fixture.componentInstance.form =
-            new FormGroup({'food': food, 'drink': new FormControl('')});
-        fixture.detectChanges();
-
-        const newForm = new FormGroup({'food': food, 'drink': new FormControl('')});
-        fixture.componentInstance.form = newForm;
-        fixture.detectChanges();
-
-        newForm.setValue({food: 'fish', drink: ''});
-        fixture.detectChanges();
-        const inputs = fixture.debugElement.queryAll(By.css('input'));
-        expect(inputs[0].nativeElement.checked).toBe(false);
-        expect(inputs[1].nativeElement.checked).toBe(true);
       });
 
       it('should update nested form group model when UI changes', () => {
@@ -338,95 +321,6 @@ export function main() {
 
       });
 
-    });
-
-    describe('select controls', () => {
-      it(`should support primitive values`, () => {
-        const fixture = initTest(FormControlNameSelect);
-        fixture.detectChanges();
-
-        // model -> view
-        const select = fixture.debugElement.query(By.css('select'));
-        const sfOption = fixture.debugElement.query(By.css('option'));
-        expect(select.nativeElement.value).toEqual('SF');
-        expect(sfOption.nativeElement.selected).toBe(true);
-
-        select.nativeElement.value = 'NY';
-        dispatchEvent(select.nativeElement, 'change');
-        fixture.detectChanges();
-
-        // view -> model
-        expect(sfOption.nativeElement.selected).toBe(false);
-        expect(fixture.componentInstance.form.value).toEqual({'city': 'NY'});
-      });
-
-      it(`should support objects`, () => {
-        const fixture = initTest(FormControlSelectNgValue);
-        fixture.detectChanges();
-
-        // model -> view
-        const select = fixture.debugElement.query(By.css('select'));
-        const sfOption = fixture.debugElement.query(By.css('option'));
-        expect(select.nativeElement.value).toEqual('0: Object');
-        expect(sfOption.nativeElement.selected).toBe(true);
-      });
-
-      it('should throw an error if compareWith is not a function', () => {
-        const fixture = initTest(FormControlSelectWithCompareFn);
-        fixture.componentInstance.compareFn = null !;
-        expect(() => fixture.detectChanges())
-            .toThrowError(/compareWith must be a function, but received null/);
-      });
-
-      it('should compare options using provided compareWith function', () => {
-        const fixture = initTest(FormControlSelectWithCompareFn);
-        fixture.detectChanges();
-
-        const select = fixture.debugElement.query(By.css('select'));
-        const sfOption = fixture.debugElement.query(By.css('option'));
-        expect(select.nativeElement.value).toEqual('0: Object');
-        expect(sfOption.nativeElement.selected).toBe(true);
-      });
-    });
-
-    describe('select multiple controls', () => {
-      it('should support primitive values', () => {
-        const fixture = initTest(FormControlSelectMultiple);
-        fixture.detectChanges();
-
-        const select = fixture.debugElement.query(By.css('select'));
-        const sfOption = fixture.debugElement.query(By.css('option'));
-        expect(select.nativeElement.value).toEqual(`0: 'SF'`);
-        expect(sfOption.nativeElement.selected).toBe(true);
-      });
-
-      it('should support objects', () => {
-        const fixture = initTest(FormControlSelectMultipleNgValue);
-        fixture.detectChanges();
-
-        const select = fixture.debugElement.query(By.css('select'));
-        const sfOption = fixture.debugElement.query(By.css('option'));
-        expect(select.nativeElement.value).toEqual('0: Object');
-        expect(sfOption.nativeElement.selected).toBe(true);
-      });
-
-      it('should throw an error when compareWith is not a function', () => {
-        const fixture = initTest(FormControlSelectMultipleWithCompareFn);
-        fixture.componentInstance.compareFn = null !;
-        expect(() => fixture.detectChanges())
-            .toThrowError(/compareWith must be a function, but received null/);
-      });
-
-      it('should compare options using provided compareWith function', fakeAsync(() => {
-           const fixture = initTest(FormControlSelectMultipleWithCompareFn);
-           fixture.detectChanges();
-           tick();
-
-           const select = fixture.debugElement.query(By.css('select'));
-           const sfOption = fixture.debugElement.query(By.css('option'));
-           expect(select.nativeElement.value).toEqual('0: Object');
-           expect(sfOption.nativeElement.selected).toBe(true);
-         }));
     });
 
     describe('form arrays', () => {
@@ -833,482 +727,6 @@ export function main() {
         fixture.detectChanges();
 
         expect(sortedClassList(formEl)).toEqual(['ng-dirty', 'ng-touched', 'ng-valid']);
-      });
-
-    });
-
-    describe('value accessors', () => {
-
-      it('should support <input> without type', () => {
-        TestBed.overrideComponent(
-            FormControlComp, {set: {template: `<input [formControl]="control">`}});
-        const fixture = initTest(FormControlComp);
-        const control = new FormControl('old');
-        fixture.componentInstance.control = control;
-        fixture.detectChanges();
-
-        // model -> view
-        const input = fixture.debugElement.query(By.css('input'));
-        expect(input.nativeElement.value).toEqual('old');
-
-        input.nativeElement.value = 'new';
-        dispatchEvent(input.nativeElement, 'input');
-
-        // view -> model
-        expect(control.value).toEqual('new');
-      });
-
-      it('should support <input type=text>', () => {
-        const fixture = initTest(FormGroupComp);
-        const form = new FormGroup({'login': new FormControl('old')});
-        fixture.componentInstance.form = form;
-        fixture.detectChanges();
-
-        // model -> view
-        const input = fixture.debugElement.query(By.css('input'));
-        expect(input.nativeElement.value).toEqual('old');
-
-        input.nativeElement.value = 'new';
-        dispatchEvent(input.nativeElement, 'input');
-
-        // view -> model
-        expect(form.value).toEqual({'login': 'new'});
-      });
-
-      it('should ignore the change event for <input type=text>', () => {
-        const fixture = initTest(FormGroupComp);
-        const form = new FormGroup({'login': new FormControl('oldValue')});
-        fixture.componentInstance.form = form;
-        fixture.detectChanges();
-
-        const input = fixture.debugElement.query(By.css('input'));
-        form.valueChanges.subscribe({next: (value) => { throw 'Should not happen'; }});
-        input.nativeElement.value = 'updatedValue';
-
-        dispatchEvent(input.nativeElement, 'change');
-      });
-
-      it('should support <textarea>', () => {
-        TestBed.overrideComponent(
-            FormControlComp, {set: {template: `<textarea [formControl]="control"></textarea>`}});
-        const fixture = initTest(FormControlComp);
-        const control = new FormControl('old');
-        fixture.componentInstance.control = control;
-        fixture.detectChanges();
-
-        // model -> view
-        const textarea = fixture.debugElement.query(By.css('textarea'));
-        expect(textarea.nativeElement.value).toEqual('old');
-
-        textarea.nativeElement.value = 'new';
-        dispatchEvent(textarea.nativeElement, 'input');
-
-        // view -> model
-        expect(control.value).toEqual('new');
-      });
-
-      it('should support <type=checkbox>', () => {
-        TestBed.overrideComponent(
-            FormControlComp, {set: {template: `<input type="checkbox" [formControl]="control">`}});
-        const fixture = initTest(FormControlComp);
-        const control = new FormControl(true);
-        fixture.componentInstance.control = control;
-        fixture.detectChanges();
-
-        // model -> view
-        const input = fixture.debugElement.query(By.css('input'));
-        expect(input.nativeElement.checked).toBe(true);
-
-        input.nativeElement.checked = false;
-        dispatchEvent(input.nativeElement, 'change');
-
-        // view -> model
-        expect(control.value).toBe(false);
-      });
-
-      describe('should support <type=number>', () => {
-        it('with basic use case', () => {
-          const fixture = initTest(FormControlNumberInput);
-          const control = new FormControl(10);
-          fixture.componentInstance.control = control;
-          fixture.detectChanges();
-
-          // model -> view
-          const input = fixture.debugElement.query(By.css('input'));
-          expect(input.nativeElement.value).toEqual('10');
-
-          input.nativeElement.value = '20';
-          dispatchEvent(input.nativeElement, 'input');
-
-          // view -> model
-          expect(control.value).toEqual(20);
-        });
-
-        it('when value is cleared in the UI', () => {
-          const fixture = initTest(FormControlNumberInput);
-          const control = new FormControl(10, Validators.required);
-          fixture.componentInstance.control = control;
-          fixture.detectChanges();
-
-          const input = fixture.debugElement.query(By.css('input'));
-          input.nativeElement.value = '';
-          dispatchEvent(input.nativeElement, 'input');
-
-          expect(control.valid).toBe(false);
-          expect(control.value).toEqual(null);
-
-          input.nativeElement.value = '0';
-          dispatchEvent(input.nativeElement, 'input');
-
-          expect(control.valid).toBe(true);
-          expect(control.value).toEqual(0);
-        });
-
-        it('when value is cleared programmatically', () => {
-          const fixture = initTest(FormControlNumberInput);
-          const control = new FormControl(10);
-          fixture.componentInstance.control = control;
-          fixture.detectChanges();
-
-          control.setValue(null);
-
-          const input = fixture.debugElement.query(By.css('input'));
-          expect(input.nativeElement.value).toEqual('');
-        });
-      });
-
-      describe('should support <type=radio>', () => {
-
-        it('should support basic functionality', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const form =
-              new FormGroup({'food': new FormControl('fish'), 'drink': new FormControl('sprite')});
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          // model -> view
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-          expect(inputs[1].nativeElement.checked).toEqual(true);
-
-          dispatchEvent(inputs[0].nativeElement, 'change');
-          fixture.detectChanges();
-
-          // view -> model
-          expect(form.get('food') !.value).toEqual('chicken');
-          expect(inputs[1].nativeElement.checked).toEqual(false);
-
-          form.get('food') !.setValue('fish');
-          fixture.detectChanges();
-
-          // programmatic change -> view
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-          expect(inputs[1].nativeElement.checked).toEqual(true);
-        });
-
-        it('should support an initial undefined value', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const form = new FormGroup({'food': new FormControl(), 'drink': new FormControl()});
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-          expect(inputs[1].nativeElement.checked).toEqual(false);
-        });
-
-        it('should reset properly', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const form =
-              new FormGroup({'food': new FormControl('fish'), 'drink': new FormControl('sprite')});
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          form.reset();
-          fixture.detectChanges();
-
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-          expect(inputs[1].nativeElement.checked).toEqual(false);
-        });
-
-        it('should set value to null and undefined properly', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const form = new FormGroup(
-              {'food': new FormControl('chicken'), 'drink': new FormControl('sprite')});
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          form.get('food') !.setValue(null);
-          fixture.detectChanges();
-
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-
-          form.get('food') !.setValue('chicken');
-          fixture.detectChanges();
-
-          form.get('food') !.setValue(undefined);
-          fixture.detectChanges();
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-        });
-
-        it('should use formControlName to group radio buttons when name is absent', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const foodCtrl = new FormControl('fish');
-          const drinkCtrl = new FormControl('sprite');
-          fixture.componentInstance.form = new FormGroup({'food': foodCtrl, 'drink': drinkCtrl});
-          fixture.detectChanges();
-
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-          expect(inputs[1].nativeElement.checked).toEqual(true);
-          expect(inputs[2].nativeElement.checked).toEqual(false);
-          expect(inputs[3].nativeElement.checked).toEqual(true);
-
-          dispatchEvent(inputs[0].nativeElement, 'change');
-          inputs[0].nativeElement.checked = true;
-          fixture.detectChanges();
-
-          const value = fixture.componentInstance.form.value;
-          expect(value.food).toEqual('chicken');
-          expect(inputs[1].nativeElement.checked).toEqual(false);
-          expect(inputs[2].nativeElement.checked).toEqual(false);
-          expect(inputs[3].nativeElement.checked).toEqual(true);
-
-          drinkCtrl.setValue('cola');
-          fixture.detectChanges();
-
-          expect(inputs[0].nativeElement.checked).toEqual(true);
-          expect(inputs[1].nativeElement.checked).toEqual(false);
-          expect(inputs[2].nativeElement.checked).toEqual(true);
-          expect(inputs[3].nativeElement.checked).toEqual(false);
-        });
-
-        it('should support removing controls from <type=radio>', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const showRadio = new FormControl('yes');
-          const form =
-              new FormGroup({'food': new FormControl('fish'), 'drink': new FormControl('sprite')});
-          fixture.componentInstance.form = form;
-          fixture.componentInstance.showRadio = showRadio;
-          showRadio.valueChanges.subscribe((change) => {
-            (change === 'yes') ? form.addControl('food', new FormControl('fish')) :
-                                 form.removeControl('food');
-          });
-          fixture.detectChanges();
-
-          const input = fixture.debugElement.query(By.css('[value="no"]'));
-          dispatchEvent(input.nativeElement, 'change');
-
-          fixture.detectChanges();
-          expect(form.value).toEqual({drink: 'sprite'});
-        });
-
-        it('should differentiate controls on different levels with the same name', () => {
-          TestBed.overrideComponent(FormControlRadioButtons, {
-            set: {
-              template: `
-              <div [formGroup]="form">
-                <input type="radio" formControlName="food" value="chicken">
-                <input type="radio" formControlName="food" value="fish">
-                <div formGroupName="nested">
-                  <input type="radio" formControlName="food" value="chicken">
-                  <input type="radio" formControlName="food" value="fish">
-                </div>
-              </div>
-              `
-            }
-          });
-          const fixture = initTest(FormControlRadioButtons);
-          const form = new FormGroup({
-            food: new FormControl('fish'),
-            nested: new FormGroup({food: new FormControl('fish')})
-          });
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          // model -> view
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.checked).toEqual(false);
-          expect(inputs[1].nativeElement.checked).toEqual(true);
-          expect(inputs[2].nativeElement.checked).toEqual(false);
-          expect(inputs[3].nativeElement.checked).toEqual(true);
-
-          dispatchEvent(inputs[0].nativeElement, 'change');
-          fixture.detectChanges();
-
-          // view -> model
-          expect(form.get('food') !.value).toEqual('chicken');
-          expect(form.get('nested.food') !.value).toEqual('fish');
-
-          expect(inputs[1].nativeElement.checked).toEqual(false);
-          expect(inputs[2].nativeElement.checked).toEqual(false);
-          expect(inputs[3].nativeElement.checked).toEqual(true);
-
-        });
-
-        it('should disable all radio buttons when disable() is called', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const form =
-              new FormGroup({food: new FormControl('fish'), drink: new FormControl('cola')});
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.disabled).toEqual(false);
-          expect(inputs[1].nativeElement.disabled).toEqual(false);
-          expect(inputs[2].nativeElement.disabled).toEqual(false);
-          expect(inputs[3].nativeElement.disabled).toEqual(false);
-
-          form.get('food') !.disable();
-          expect(inputs[0].nativeElement.disabled).toEqual(true);
-          expect(inputs[1].nativeElement.disabled).toEqual(true);
-          expect(inputs[2].nativeElement.disabled).toEqual(false);
-          expect(inputs[3].nativeElement.disabled).toEqual(false);
-
-          form.disable();
-          expect(inputs[0].nativeElement.disabled).toEqual(true);
-          expect(inputs[1].nativeElement.disabled).toEqual(true);
-          expect(inputs[2].nativeElement.disabled).toEqual(true);
-          expect(inputs[3].nativeElement.disabled).toEqual(true);
-
-          form.enable();
-          expect(inputs[0].nativeElement.disabled).toEqual(false);
-          expect(inputs[1].nativeElement.disabled).toEqual(false);
-          expect(inputs[2].nativeElement.disabled).toEqual(false);
-          expect(inputs[3].nativeElement.disabled).toEqual(false);
-        });
-
-        it('should disable all radio buttons when initially disabled', () => {
-          const fixture = initTest(FormControlRadioButtons);
-          const form = new FormGroup({
-            food: new FormControl({value: 'fish', disabled: true}),
-            drink: new FormControl('cola')
-          });
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          const inputs = fixture.debugElement.queryAll(By.css('input'));
-          expect(inputs[0].nativeElement.disabled).toEqual(true);
-          expect(inputs[1].nativeElement.disabled).toEqual(true);
-          expect(inputs[2].nativeElement.disabled).toEqual(false);
-          expect(inputs[3].nativeElement.disabled).toEqual(false);
-        });
-
-      });
-
-      describe('should support <type=range>', () => {
-        it('with basic use case', () => {
-          const fixture = initTest(FormControlRangeInput);
-          const control = new FormControl(10);
-          fixture.componentInstance.control = control;
-          fixture.detectChanges();
-
-          // model -> view
-          const input = fixture.debugElement.query(By.css('input'));
-          expect(input.nativeElement.value).toEqual('10');
-
-          input.nativeElement.value = '20';
-          dispatchEvent(input.nativeElement, 'input');
-
-          // view -> model
-          expect(control.value).toEqual(20);
-        });
-
-        it('when value is cleared in the UI', () => {
-          const fixture = initTest(FormControlNumberInput);
-          const control = new FormControl(10, Validators.required);
-          fixture.componentInstance.control = control;
-          fixture.detectChanges();
-
-          const input = fixture.debugElement.query(By.css('input'));
-          input.nativeElement.value = '';
-          dispatchEvent(input.nativeElement, 'input');
-
-          expect(control.valid).toBe(false);
-          expect(control.value).toEqual(null);
-
-          input.nativeElement.value = '0';
-          dispatchEvent(input.nativeElement, 'input');
-
-          expect(control.valid).toBe(true);
-          expect(control.value).toEqual(0);
-        });
-
-        it('when value is cleared programmatically', () => {
-          const fixture = initTest(FormControlNumberInput);
-          const control = new FormControl(10);
-          fixture.componentInstance.control = control;
-          fixture.detectChanges();
-
-          control.setValue(null);
-
-          const input = fixture.debugElement.query(By.css('input'));
-          expect(input.nativeElement.value).toEqual('');
-        });
-      });
-
-      describe('custom value accessors', () => {
-        it('should support basic functionality', () => {
-          const fixture = initTest(WrappedValueForm, WrappedValue);
-          const form = new FormGroup({'login': new FormControl('aa')});
-          fixture.componentInstance.form = form;
-          fixture.detectChanges();
-
-          // model -> view
-          const input = fixture.debugElement.query(By.css('input'));
-          expect(input.nativeElement.value).toEqual('!aa!');
-
-          input.nativeElement.value = '!bb!';
-          dispatchEvent(input.nativeElement, 'input');
-
-          // view -> model
-          expect(form.value).toEqual({'login': 'bb'});
-
-          // custom validator
-          expect(form.get('login') !.errors).toEqual({'err': true});
-          form.setValue({login: 'expected'});
-          expect(form.get('login') !.errors).toEqual(null);
-        });
-
-        it('should support non builtin input elements that fire a change event without a \'target\' property',
-           () => {
-             const fixture = initTest(MyInputForm, MyInput);
-             fixture.componentInstance.form = new FormGroup({'login': new FormControl('aa')});
-             fixture.detectChanges();
-
-             const input = fixture.debugElement.query(By.css('my-input'));
-             expect(input.componentInstance.value).toEqual('!aa!');
-
-             input.componentInstance.value = '!bb!';
-             input.componentInstance.onInput.subscribe((value: any) => {
-               expect(fixture.componentInstance.form.value).toEqual({'login': 'bb'});
-             });
-             input.componentInstance.dispatchChangeEvent();
-           });
-
-        it('should support custom accessors without setDisabledState - formControlName', () => {
-          const fixture = initTest(WrappedValueForm, WrappedValue);
-          fixture.componentInstance.form = new FormGroup({
-            'login': new FormControl({value: 'aa', disabled: true}),
-          });
-          fixture.detectChanges();
-          expect(fixture.componentInstance.form.status).toEqual('DISABLED');
-          expect(fixture.componentInstance.form.get('login') !.status).toEqual('DISABLED');
-        });
-
-        it('should support custom accessors without setDisabledState - formControlDirective',
-           () => {
-             TestBed.overrideComponent(
-                 FormControlComp,
-                 {set: {template: `<input type="text" [formControl]="control" wrapped-value>`}});
-             const fixture = initTest(FormControlComp);
-             fixture.componentInstance.control = new FormControl({value: 'aa', disabled: true});
-             fixture.detectChanges();
-             expect(fixture.componentInstance.control.status).toEqual('DISABLED');
-           });
-
       });
 
     });
@@ -1992,44 +1410,6 @@ export function main() {
   });
 }
 
-@Directive({
-  selector: '[wrapped-value]',
-  host: {'(input)': 'handleOnInput($event.target.value)', '[value]': 'value'},
-  providers: [
-    {provide: NG_VALUE_ACCESSOR, multi: true, useExisting: WrappedValue},
-    {provide: NG_VALIDATORS, multi: true, useExisting: WrappedValue}
-  ]
-})
-class WrappedValue implements ControlValueAccessor {
-  value: any;
-  onChange: Function;
-
-  writeValue(value: any) { this.value = `!${value}!`; }
-
-  registerOnChange(fn: (value: any) => void) { this.onChange = fn; }
-  registerOnTouched(fn: any) {}
-
-  handleOnInput(value: any) { this.onChange(value.substring(1, value.length - 1)); }
-
-  validate(c: AbstractControl) { return c.value === 'expected' ? null : {'err': true}; }
-}
-
-@Component({selector: 'my-input', template: ''})
-class MyInput implements ControlValueAccessor {
-  @Output('input') onInput = new EventEmitter();
-  value: string;
-
-  constructor(cd: NgControl) { cd.valueAccessor = this; }
-
-  writeValue(value: any) { this.value = `!${value}!`; }
-
-  registerOnChange(fn: (value: any) => void) { this.onInput.subscribe({next: fn}); }
-
-  registerOnTouched(fn: any) {}
-
-  dispatchChangeEvent() { this.onInput.emit(this.value.substring(1, this.value.length - 1)); }
-}
-
 function uniqLoginAsyncValidator(expectedValue: string, timeout: number = 0) {
   return (c: AbstractControl) => {
     let resolve: (result: any) => void;
@@ -2110,39 +1490,6 @@ class NestedFormGroupComp {
 }
 
 @Component({
-  selector: 'form-control-number-input',
-  template: `<input type="number" [formControl]="control">`
-})
-class FormControlNumberInput {
-  control: FormControl;
-}
-
-@Component({
-  selector: 'form-control-range-input',
-  template: `<input type="range" [formControl]="control">`
-})
-class FormControlRangeInput {
-  control: FormControl;
-}
-
-@Component({
-  selector: 'form-control-radio-buttons',
-  template: `
-  <form [formGroup]="form" *ngIf="showRadio.value === 'yes'">
-    <input type="radio" formControlName="food" value="chicken">
-    <input type="radio" formControlName="food" value="fish">
-    <input type="radio" formControlName="drink" value="cola">
-    <input type="radio" formControlName="drink" value="sprite">
-  </form>
-  <input type="radio" [formControl]="showRadio" value="yes">
-  <input type="radio" [formControl]="showRadio" value="no">`
-})
-class FormControlRadioButtons {
-  form: FormGroup;
-  showRadio = new FormControl('yes');
-}
-
-@Component({
   selector: 'form-array-comp',
   template: `
     <div [formGroup]="form">
@@ -2175,115 +1522,6 @@ class FormArrayNestedGroup {
   cityArray: FormArray;
 }
 
-@Component({
-  selector: 'form-control-name-select',
-  template: `
-    <div [formGroup]="form">
-      <select formControlName="city">
-        <option *ngFor="let c of cities" [value]="c"></option>
-      </select>
-    </div>`
-})
-class FormControlNameSelect {
-  cities = ['SF', 'NY'];
-  form = new FormGroup({city: new FormControl('SF')});
-}
-
-@Component({
-  selector: 'form-control-select-ngValue',
-  template: `
-    <div [formGroup]="form">
-      <select formControlName="city">
-        <option *ngFor="let c of cities" [ngValue]="c">{{c.name}}</option>
-      </select>
-    </div>`
-})
-class FormControlSelectNgValue {
-  cities = [{id: 1, name: 'SF'}, {id: 2, name: 'NY'}];
-  form = new FormGroup({city: new FormControl(this.cities[0])});
-}
-
-@Component({
-  selector: 'form-control-select-compare-with',
-  template: `
-    <div [formGroup]="form">
-      <select formControlName="city" [compareWith]="compareFn">
-        <option *ngFor="let c of cities" [ngValue]="c">{{c.name}}</option>
-      </select>
-    </div>`
-})
-class FormControlSelectWithCompareFn {
-  compareFn:
-      (o1: any, o2: any) => boolean = (o1: any, o2: any) => o1 && o2? o1.id === o2.id: o1 === o2;
-  cities = [{id: 1, name: 'SF'}, {id: 2, name: 'NY'}];
-  form = new FormGroup({city: new FormControl({id: 1, name: 'SF'})});
-}
-
-@Component({
-  selector: 'form-control-select-multiple',
-  template: `
-    <div [formGroup]="form">
-      <select multiple formControlName="city">
-        <option *ngFor="let c of cities" [value]="c">{{c}}</option>
-      </select>
-    </div>`
-})
-class FormControlSelectMultiple {
-  cities = ['SF', 'NY'];
-  form = new FormGroup({city: new FormControl(['SF'])});
-}
-
-@Component({
-  selector: 'form-control-select-multiple',
-  template: `
-    <div [formGroup]="form">
-      <select multiple formControlName="city">
-        <option *ngFor="let c of cities" [ngValue]="c">{{c.name}}</option>
-      </select>
-    </div>`
-})
-class FormControlSelectMultipleNgValue {
-  cities = [{id: 1, name: 'SF'}, {id: 2, name: 'NY'}];
-  form = new FormGroup({city: new FormControl([this.cities[0]])});
-}
-
-@Component({
-  selector: 'form-control-select-multiple-compare-with',
-  template: `
-    <div [formGroup]="form">
-      <select multiple formControlName="city" [compareWith]="compareFn">
-        <option *ngFor="let c of cities" [ngValue]="c">{{c.name}}</option>
-      </select>
-    </div>`
-})
-class FormControlSelectMultipleWithCompareFn {
-  compareFn:
-      (o1: any, o2: any) => boolean = (o1: any, o2: any) => o1 && o2? o1.id === o2.id: o1 === o2;
-  cities = [{id: 1, name: 'SF'}, {id: 2, name: 'NY'}];
-  form = new FormGroup({city: new FormControl([{id: 1, name: 'SF'}])});
-}
-
-@Component({
-  selector: 'wrapped-value-form',
-  template: `
-   <div [formGroup]="form">
-    <input type="text" formControlName="login" wrapped-value>
-  </div>`
-})
-class WrappedValueForm {
-  form: FormGroup;
-}
-
-@Component({
-  selector: 'my-input-form',
-  template: `
-   <div [formGroup]="form">
-      <my-input formControlName="login"></my-input>
-   </div>`
-})
-class MyInputForm {
-  form: FormGroup;
-}
 
 @Component({
   selector: 'form-group-ng-model',
